Remove API transformers
